requirements.txt
aptdaemon==2.0.2
asgiref==3.8.1
attrs==25.3.0
autocommand==2.2.2
babel==2.17.0
bcc==0.31.0
bcrypt==4.2.0
beautifulsoup4==4.13.4
blinker==1.9.0
Brlapi==0.8.6
certifi==2025.1.31
chardet==5.2.0
click==8.1.8
command-not-found==0.3
configobj==5.0.9
cryptography==43.0.0
cssselect==1.3.0
cupshelpers==1.0
dasbus==1.7
dbus-python==1.4.0
defer==1.0.6
distro==1.9.0
distro-info==1.14
duplicity==3.0.4
fasteners==0.18
Flask==3.1.2
gyp-next==0.16.2
html5lib-modern==1.2
httplib2==0.22.0
idna==3.10
inflect==7.3.1
itsdangerous==2.2.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
jsonpatch==1.32
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.12.1
language-selector==0.1
launchpadlib==2.1.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
linkify-it-py==2.0.3
louis==3.33.0
lxml==6.0.1
Mako==1.3.9.dev0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
more-itertools==10.7.0
netaddr==1.3.0
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.47
packaging==25.0
paramiko==3.5.1
passlib==1.7.4
pexpect==4.9.0
pillow==11.3.0
psutil==7.0.0
ptyprocess==0.7.0
pyasyncore==1.0.2
pycairo==1.27.0
pycups==2.0.4
Pygments==2.18.0
PyGObject==3.50.0
pyinotify==0.9.6
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pyparsing==3.1.2
pyserial==3.5
python-apt==3.0.0+ubuntu1.1
python-dateutil==2.9.0
python-debian==1.0.1+ubuntu1
pyxdg==0.28
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
setproctitle==1.3.6
soupsieve==2.7
speedtest-cli==2.1.3
systemd-python==235
typeguard==4.4.2
typing_extensions==4.13.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
uc-micro-py==1.0.3
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.3.0
usb-creator==0.3.16
wadllib==2.0.0
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.46.1
xdg==5
xkit==0.0.0
